requirements.txt
accelerate==0.30.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
attrs==23.2.0
base58==2.1.1
beautifulsoup4==4.12.3
black==21.10b0
blinker==1.4
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
# command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
# dblatex==0.3.12
dbus-python==1.2.18
diffusers==0.24.0
distlib==0.3.4
distro==1.7.0
# distro-info==1.1+ubuntu0.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
Faker==25.4.0
favicon==0.7.0
filelock==3.6.0
flake8==4.0.1
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib-metadata==4.6.4
install==1.3.5
isort==5.6.4
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llvmlite==0.42.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.6.1
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
mypy-extensions==0.4.3
netifaces==0.11.0
# networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
onnxruntime==1.18.0
openai==1.30.5
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==21.3
pandas==2.2.1
pathspec==0.9.0
pillow==10.2.0
platformdirs==2.5.1
pluggy==0.13.0
pooch==1.8.1
prometheus_client==0.20.0
protobuf==3.20.3
psutil==5.9.8
py==1.10.0
pyarrow==16.1.0
pycodestyle==2.8.0
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
pyflakes==2.4.0
Pygments==2.18.0
# PyGObject==3.42.1
PyJWT==2.3.0
PyMatting==1.1.12
pymdown-extensions==10.8.1
pyparsing==2.4.7
# python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
pytz==2024.1
# PyYAML==5.4.1
referencing==0.35.1
regex==2024.5.15
rembg==2.0.57
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
# scikit-image==0.23.2
scipy==1.13.1
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.1
st-pages==0.4.5
st-theme==1.2.3
streamlit==1.35.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-cropper==0.2.0
streamlit-cropperjs==0.0.9
streamlit-drawable-canvas==0.9.3
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-image-select==0.6.0
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12.1
# systemd-python==234
tenacity==8.3.0
tifffile==2024.5.22
tokenizers==0.19.1
toml==0.10.2
tomli==1.2.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tox==3.21.4
tqdm==4.66.4
transformers==4.41.2
triton==2.3.0
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
urllib3==2.2.1
validators==0.28.3
# virtualenv==20.13.0+ds
wadllib==1.3.6
watchdog==4.0.1
zipp==1.0.0
